fix NPE in TableViewer [will be separately merged]
Changing the number of visible columns 100 times triggers a bug [1] in
OpenJFX, so we have to work around by never removing columns from the
table.

Never hide Columns, just reduce width, so the NPE is not triggered
Only update table when it is visible, reduces load, but the less code is
running, the less bugs can occur.
Refresh the selection model on change of elements, unfortunately this
resets the selction.

[1] https://bugs.openjdk.java.net/browse/JDK-8217953